Update lightning page and readme and added test class
Updated the lightning page with full button functionality; only thing it doesn't currently do is reload the page after it's changed the status.  (Modal box could be more attractive as well.)  Added a test class to check the Apex controller but haven't coded it yet.  Updated readme to start explaining the data factory.